Revert to release 1.0
