+++ README.md
@@ -1,148 +1,9 @@
-[![Check for prettier](https://github.com/JaydenDev/freecat/actions/workflows/checks.yml/badge.svg?branch=main)](https://github.com/JaydenDev/freecat/actions/workflows/checks.yml)
-![JavaScript](https://img.shields.io/badge/javascript-%23323330.svg?style=for-the-badge&logo=javascript&logoColor=%23F7DF1E)
-![HTML5](https://img.shields.io/badge/html5-%23E34F26.svg?style=for-the-badge&logo=html5&logoColor=white)
-![CSS3](https://img.shields.io/badge/css3-%231572B6.svg?style=for-the-badge&logo=css3&logoColor=white)
-![Electron.js](https://img.shields.io/badge/Electron-191970?style=for-the-badge&logo=Electron&logoColor=white)
-![License](https://camo.githubusercontent.com/3dbcfa4997505c80ef928681b291d33ecfac2dabf563eb742bb3e269a5af909c/68747470733a2f2f696d672e736869656c64732e696f2f6769746875622f6c6963656e73652f496c65726961796f2f6d61726b646f776e2d6261646765733f7374796c653d666f722d7468652d6261646765)
-
 # The FreeCat Browser
-
-The user interface is entirely rendered as regular HTML, backed by Electron
-
-![image](https://user-images.githubusercontent.com/92550746/145330664-a32ab839-84e4-4a27-9581-25d3248b8b4e.png)
-
+Hello
 ## Installation
-
 ```bash
-git clone https://github.com/JaydenDev/freecat/
-cd freecat
+cd app
 npm install
+cd ..
+./run-linux.sh (or the script depending on platform)
 ```
-
-## Running
-
-`npm start`
-
-# Builds
-
-[Download Here](https://dl.jaydendev.repl.co/freecat)
-Or check the releases tab.
-
-# Constant Channel
-
-Test the rolling constant channel by doing:
-
-```bash
-git clone https://github.com/JaydenDev/freecat
-```
-
-# Contributors
-
-<table>
-<tr>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
-=======
->>>>>>> parent of 71385e9 (Unstable (#81))
-
-        <a href=https://github.com/hello-smile6>
-            <img src=https://avatars.githubusercontent.com/u/73048226?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=hello-smile6/>
-            <br />
-            <sub style="font-size:14px"><b>hello-smile6</b></sub>
-        </a>
-    </td>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/JaydenDev>
-            <img src=https://avatars.githubusercontent.com/u/92550746?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=JaydenDev/>
-            <br />
-            <sub style="font-size:14px"><b>JaydenDev</b></sub>
-
-> > > > > > > parent of 71385e9 (Unstable (#81))
-
-        <a href=https://github.com/JaydenDev>
-            <img src=https://avatars.githubusercontent.com/u/92550746?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=JaydenDev/>
-            <br />
-            <sub style="font-size:14px"><b>JaydenDev</b></sub>
-        </a>
-    </td>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/hello-smile6>
-            <img src=https://avatars.githubusercontent.com/u/73048226?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=hello-smile6/>
-            <br />
-            <sub style="font-size:14px"><b>hello-smile6</b></sub>
-        </a>
-    </td>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/webdev03>
-            <img src=https://avatars.githubusercontent.com/u/75148774?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=webdev03/>
-            <br />
-            <sub style="font-size:14px"><b>webdev03</b></sub>
-        </a>
-    </td>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/ninjamar>
-            <img src=https://avatars.githubusercontent.com/u/70116773?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=ninjamar/>
-            <br />
-            <sub style="font-size:14px"><b>ninjamar</b></sub>
-        </a>
-    </td>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/ScolderCreations>
-            <img src=https://avatars.githubusercontent.com/u/69083943?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=Scolder/>
-            <br />
-            <sub style="font-size:14px"><b>Scolder</b></sub>
-        </a>
-    </td>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/MystPi>
-            <img src=https://avatars.githubusercontent.com/u/86574651?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=MystPi/>
-            <br />
-            <sub style="font-size:14px"><b>MystPi</b></sub>
-        </a>
-    </td>
-
-</tr>
-<tr>
-    <td align="center" style="word-wrap: break-word; width: 150.0; height: 150.0">
-        <a href=https://github.com/kccuber-scratch>
-            <img src=https://avatars.githubusercontent.com/u/92891642?v=4 width="100;"  style="border-radius:50%;align-items:center;justify-content:center;overflow:hidden;padding-top:10px" alt=kccuber/>
-            <br />
-            <sub style="font-size:14px"><b>kccuber</b></sub>
-        </a>
-    </td>
-</tr>
-</table>
-
-# Contributing
-
-### Don't:
-
-Malicious Code \
-Messing around with PRs
-
-### Do:
-
-Name PRs properly \
-Test PRs before asking for merging
-
-# Collaborators
-
-### Don't:
-
-Abuse your powers \
-Make unwanted changes \
-Make changes without permission
-
-### Do:
-
-Ask before merging PRs
-
-# Clean your code
-
-First, run `npm run check`. To confirm that formatting is needed. Then run `npm run format` in your shell to format your code. It will automatically commit, and your clean sleek code will be submitted.
-
-# Release Progress
-
-![progress](https://progress-bar.dev/0/?title=2.0)
