+++ app/config/userchrome.css
@@ -1,0 +1,37 @@
+body {
+	background-color: white;
+}
+
+webview {
+	width: 100%;
+	height: 90%;
+	overflow-x: hidden;
+	overflow-y: hidden;
+}
+
+* {
+	font-family: monospace;
+	color: blue;
+	font-size: 15pt;
+	list-style-type: none;
+	border: none;
+	border-width: 0px;
+}
+
+input {
+	padding: 4pt;
+	background-color: white;
+	color: black;
+	width: 100%;
+	outline-color: #aa00ff;
+}
+
+button {
+	background-color: white;
+	color: black;
+	width: 100%;
+}
+td, tr {
+	background-color: #0e0e0e;
+	color: white
+}
