+++ app/home.html
@@ -1,0 +1,19 @@
+<!DOCTYPE HTML>
+<HTML>
+	<HEAD>
+		<script src="js/searchbox.js"></script>
+		<link rel="stylesheet" href="config/userchrome.css">
+	</HEAD>
+	<BODY>
+		<h1> 🐈  FreeCat! </h1>
+		<br>
+		<p> Customize your home page by editing src/home.html </p>
+		<br>
+		<p> freecat-electron-adapted Version 21.12.420 </p>
+		<p> Entirely customizable and user configurable, under a free license. Feel free to fork! </p>
+	</BODY>
+</HTML>
+
+
+
+
