+++ app/index.html
@@ -1,0 +1,5 @@
+
+<script src="js/dynamic.js"></script>
+<link rel="stylesheet" href="config/userchrome.css">
+<input type="text" placeholder="🔎 Type to Search" id="searchbox" name="searchbox"> <button onclick="webload()">Go</button>
+<webview id="view" src="home.html"></webview>
