+++ app/js/dynamic.js
@@ -1,0 +1,4 @@
+function webload() {
+    let wsite = document.getElementById("searchbox").value;
+    document.getElementById("view").src=wsite;
+}
