+++ app/js/searchbox.js
@@ -1,0 +1,10 @@
+
+
+function redirect() {
+	const wsite = document.getElementById("searchbox").value
+	window.location.href = wsite;
+}
+
+function gotoWebLoader() {
+	window.location.href = "webload.html";
+}
