+++ app/package.json
@@ -1,15 +1,12 @@
 {
   "name": "freecat",
-  "version": "1.0.0",
+  "version": "1.0",
   "description": "Freecat Adapted Electron Application",
   "main": "main/main.js",
   "scripts": {
     "start": "electron .",
     "package": "electron-forge package",
-    "make": "electron-forge make",
-    "check": "prettier --check .",
-    "format": "prettier --write .",
-    "eslint": "eslint ."
+    "make": "electron-forge make"
   },
   "repository": "https://github.com/electron/electron-quick-start",
   "keywords": [
@@ -26,14 +23,10 @@
     "@electron-forge/maker-squirrel": "^6.0.0-beta.61",
     "@electron-forge/maker-zip": "^6.0.0-beta.61",
     "electron": "^16.0.0",
-    "electron-prebuilt-compile": "^1.8.2",
-    "eslint": "^8.4.1",
-    "eslint-config-prettier": "^8.3.0",
-    "prettier": "2.5.1"
+    "electron-prebuilt-compile": "1.8.2"
   },
   "dependencies": {
-    "electron-squirrel-startup": "^1.0.0",
-    "font-awesome": "^4.7.0"
+    "electron-squirrel-startup": "^1.0.0"
   },
   "config": {
     "forge": {
