--- config/env.css
@@ -1,3 +1,0 @@
-:root {
-  --app-url-bar-size: 20px;
-}
