--- config/userchrome.css
@@ -1,73 +1,0 @@
-@import url("./env.css");
-body,
-html {
-  background-color: white;
-  padding: 0 !important;
-  margin: 0 !important;
-  border: 0 !important;
-  width: 100%;
-  height: 100%;
-}
-nav {
-  width: 100vw;
-  height: var(--app-url-bar-size);
-  padding: 0;
-  margin: 0;
-  border: 0;
-  display: block;
-  box-sizing: border-box;
-}
-#searchbox {
-  height: var(--app-url-bar-size);
-  width: calc(100vw - var(--app-url-bar-size)) !important;
-  padding: 0 !important;
-  border: 0 !important;
-  margin: 0 !important;
-  box-sizing: border-box;
-  display: block !important;
-  float: left;
-}
-#load {
-  width: var(--app-url-bar-size) !important;
-  font-size: 12px;
-  height: var(--app-url-bar-size);
-  box-sizing: border-box;
-  display: block;
-  padding: 0 !important;
-  border: 0 !important;
-  margin: 0 !important;
-}
-webview {
-  width: 100vw;
-  height: calc(100vh - var(--app-url-bar-size));
-  box-sizing: border-box;
-  overflow-x: visible;
-  overflow-y: visible;
-}
-* {
-  font-family: sans-serif;
-  color: blue;
-  font-size: 16px;
-  list-style-type: none;
-  border: none;
-  border-width: 0px;
-}
-/* 
-input {
-	padding: 4pt;
-	background-color: white;
-	color: black;
-	width: 100%;
-	outline-color: #aa00ff;
-}
-
-button {
-	background-color: white;
-	color: black;
-	width: 100%;
-} */
-td,
-tr {
-  background-color: #0e0e0e;
-  color: white;
-}
