--- contrib/contrib.txt
@@ -1,3 +1,0 @@
-@hello-smile6 - For the improved User Interface, Bug Fixes, and logos 
-@webdev03 - For bug fixes, features, and suggesting the use of webviews. As well as creating the FreeCat logo 
-@ScolderCreations - For bug fixes and suggestions
