--- home.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <script src="js/searchbox.js"></script>
-    <link rel="stylesheet" href="config/userchrome.css" />
-  </head>
-  <body>
-    <h1>🐈 FreeCat!</h1>
-    <br />
-    <p>
-      Customize your home page by editing config/env.css and
-      config/userchrome.css
-    </p>
-    <br />
-    <p>FreeCat dev</p>
-    <p>
-      Entirely customizable and user configurable, under a free license. Feel
-      free to fork!
-    </p>
-  </body>
-</html>
