--- index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Freecat</title>
-    <link rel="stylesheet" href="config/userchrome.css" />
-    <meta http-equiv="Content-Security-Policy" content="script-src 'self'" />
-    <link href="icons/seconds-32x32.svg" rel="shortcut icon" />
-    <link
-      rel="stylesheet"
-      href="node_modules/font-awesome/css/font-awesome.min.css"
-    />
-  </head>
-  <body>
-    <nav>
-      <input
-        placeholder="Type URL"
-        id="searchbox"
-        name="searchbox"
-        type="url"
-      />
-      <button id="load"><i class="fa fa-search"></i></button>
-    </nav>
-    <webview id="view" src="home.html"></webview>
-    <script src="js/dynamic.js"></script>
-  </body>
-</html>
