--- js/searchbox.js
@@ -1,9 +1,0 @@
-/* exported redirect */
-function redirect() {
-  const wsite = document.getElementById("searchbox").value
-  window.location.href = wsite
-}
-
-function gotoWebLoader() {
-  window.location.href = "webload.html"
-}
