+++ run-linux.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd app/
+npm start
