+++ run-win.bat
@@ -1,0 +1,3 @@
+@echo off
+cd app/
+npm start
